+++ changelog.md
@@ -7,6 +7,7 @@
 * Add hardware in the loop tests.
 * Add empty APN settings window.
 * Add New/Edit APN window
+* Add APN options window
 
 ### Changed